+++ changelog.md
@@ -12,6 +12,7 @@ Codion Change Log
 - FilterTable.COLUMN_RESIZING configuration value added.
 - FilterTable.Builder.cellSelectionEnabled() renamed cellSelection().
 - JSpecify annotations added.
+- TemporalFieldPanel.getTemporal() and setTemporal() renamed get() and set().
 ### is.codion.framework.domain
 - DefaultEntity.validateTypes() performance bug fixed, unnecessary HashMap lookup removed, renamed validate().
 - DefaultKey.setSerializer() and serializerForDomain() moved to EntitySerializer.